Rename files because of file restructure
